Make one search per line in file (use-individual-search-terms)
Previously we had combined all of the search terms in the file
into one query to reduce the cost of each query. However, as the
maximum number of results per query is only 50, we have changed
the implementation to make one query per line in the file. Thus
allowing us to get more results.